// src/app/shared/models/config.ts
export class Config {
    apiUrl: string;
    apiKey: string;
    movieUrl: string;
    seacrhMovieUrl: string;
    params: string;
    imgPath: string;
    midImgPath: string;
    smallImgPath: string;
    bigBackPath: string;
    midBackPath: string;
    smallBackPath: string;
    personUrl: string;
    searchPersonUrl: string;
}